+++ progs/ballast.py
@@ -11,6 +11,7 @@ import os
 import argparse
 import typing as tp
 import configparser as cfg
+from math import ceil
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -76,9 +77,49 @@ def parse_args(args: tp.Sequence[str]) -> argparse.Namespace:
     return parser.parse_args(args)
 
 
-def parse_inifile(fname) -> tp.Tuple[tp.Dict[str, tp.Any],
-                                     SpecLayout,
-                                     tp.Dict[str, tp.Any]]:
+def parse_subid(ident: str, ncols: int = 1) -> tp.Tuple[int, int]:
+    """Parses a subplot identifier.
+
+    Takes a subplot identifier and returns the corresponding row and
+      column.
+
+    Parameters
+    ----------
+    ident
+        Identifier string.
+    ncols
+        Number of columns.
+
+    Returns
+    -------
+    int
+        Row index (starting from 1).
+    int
+        Column index (starting from 1).
+
+    Raises
+    ------
+    ValueError
+        Unable to parse the subplot specification.
+    """
+    grid = ident.split(',')
+    if len(grid) == 2:
+        row = max(int(grid[0]), 1)
+        col = max(int(grid[1]), 1)
+    elif len(grid) == 1:
+        i = int(grid)
+        row = max(int(ceil(i/ncols)), 1)
+        col = max(i - (row-1)*ncols, 1)
+    else:
+        raise ValueError('Incorrect subplot specification.')
+
+    return row, col
+
+
+def parse_inifile(fname: str
+                  ) -> tp.Tuple[tp.Dict[str, tp.Any],
+                                tp.List[tp.List[SpecLayout]],
+                                tp.Dict[str, tp.Any]]:
     """Parses INI file.
 
     Parses a INI configuration file.
@@ -86,104 +127,206 @@ def parse_inifile(fname) -> tp.Tuple[tp.Dict[str, tp.Any],
     Parameters
     ----------
     fname
-        Filename
+        Filename.
 
     Returns
     -------
-    figdat
-        Figure data
-    spcdat
-        Spectrum layout parameters
-    curves
-        Curves data
+    dict
+        Figure data.
+    list
+        List of lists of spectrum layout parameters (grif format).
+    dict
+        Curves data.
 
     Raises
     ------
     FileNotFoundError
-        INI file or input file missing
+        INI file or input file missing.
     ValueError
-        Incorrect parameter
+        Incorrect parameter.
     """
     if not os.path.exists(fname):
         raise FileNotFoundError('Missing INI file')
     opts = cfg.ConfigParser()
     opts.read(fname)
-    secs = {key.lower(): key for key in opts.sections()}
-    if 'layout' in secs:
-        optlay = opts[secs['layout']]
-        _title = optlay.get('Title', None)
-        _xleft = optlay.get('XLeft', None) or optlay.get('XMin', None)
-        _xright = optlay.get('XRight', None) or optlay.get('XMax', None)
-        _ytop = optlay.get('YTop', None) or optlay.get('YMax', None)
-        _ybottom = optlay.get('YBottom', None) or optlay.get('YMin', None)
-        _xscale = optlay.get('XScale', 'linear')
-        _yscale = optlay.get('YScale', 'linear')
-        _xlabel = optlay.get('XLabel', None)
-        _ylabel = optlay.get('YLabel', None)
-        spcdat = SpecLayout(_xleft, _xright, _ytop, _ybottom, _xscale, _yscale,
-                            _xlabel, _ylabel, _title)
-        _legpos = optlay.get('Legend', 'best')
-        _legcol = optlay.get('Legend_Cols', 1)
-        spcdat.legend(_legpos, _legcol)
-    else:
-        spcdat = SpecLayout()
+    secs = {key.strip().lower(): key for key in opts.sections()}
 
     figdat = {
+        'title': None,
         'geom': None,
+        'shareaxes': False,
+        'subp': (1, 1),
     }
     if 'figure' in secs:
-        optfig = opts[secs['figure']]
-        _geom = optfig.get('Geometry', None)
-        if _geom is not None:
-            _res = _geom.replace('(', '').replace(')', '').split(',')
-        if len(_res) == 1:
-            raise ValueError('Incorrect value for geometry.')
+        optsec = opts[secs['figure']]
+        figdat['title'] = optsec.get('maintitle', fallback=None)
+        res = optsec.get('mergeaxes', fallback='None').lower()
+        if res == 'none':
+            val = False
+        elif res == 'x':
+            val = 'X'
+        elif res == 'y':
+            val = 'Y'
+        elif res == 'all':
+            val = True
         else:
-            _geom = (float(_res[0]), float(_res[1]))
-        figdat['geom'] = _geom
+            raise ValueError('Unrecognized value for MergeAxes')
+        figdat['shareaxes'] = val
+        optkey = optsec.get('geometry', None)
+        if optkey is not None:
+            res = optkey.replace('(', '').replace(')', '').split(',')
+            if len(res) == 1:
+                raise ValueError('Incorrect value for geometry.')
+            else:
+                val = (float(res[0]), float(res[1]))
+            figdat['geom'] = val
+        optkey = optsec.get('subplots', None)
+        if optkey is not None:
+            res = optkey.replace('(', '').replace(')', '').split(',')
+            if len(res) == 1:
+                val = (max(int(res[0]), 1), 1)
+            else:
+                val = (max(int(res[0]), 1), max(int(res[1]), 1))
+        else:
+            val = None
+    else:
+        val = None
+    if val is not None:
+        figdat['subp'] = val
+    nrows, ncols = figdat['subp']
+    figdat['nums'] = nrows * ncols
+    spcdat = []
+    for _ in range(nrows):
+        spcdat.append([None for j in range(ncols)])
+
+    # The layout system works in a slightly different way than curves
+    # Besides using defaults, users can use the generic [layout] to define
+    #   a common layout.
+    # We first build some default setup, which will be used for all others.
+    # The keys correspond to SpecLayout
+    spckeys = {
+        'title': ('title', ),
+        'xleft': ('xleft', 'xmin'),
+        'xright': ('xright', 'xmax'),
+        'ytop': ('ytop', 'ymax'),
+        'ybottom': ('ybottom', 'ymin'),
+        'xscale': ('xscale', ),
+        'yscale': ('yscale', ),
+        'xlabel': ('xlabel', ),
+        'ylabel': ('ylabel', ),
+        'legpos': ('legend', ),
+        'legcol': ('legend_cols', )
+    }
+    spcbase = {
+        'title': None,
+        'xleft': None,
+        'xright': None,
+        'ytop': None,
+        'ybottom': None,
+        'xscale': 'linear',
+        'yscale': 'linear',
+        'xlabel': None,
+        'ylabel': None,
+        'legpos': 'best',
+        'legcol': 1
+    }
+    if 'layout' in secs:
+        optsec = opts[secs['layout']]
+        for key in spckeys:
+            for alias in spckeys[key]:
+                if alias in optsec:
+                    spcbase[key] = optsec[alias]
+                    break
+
+    for sec in secs:
+        if sec.startswith('layout'):
+            res = sec.split(':')
+            if len(res) == 2:  # Ignore default case here
+                row, col = parse_subid(res[1], figdat['subp'][1])
+                if row > figdat['subp'][0] or col > figdat['subp'][1]:
+                    break
+                # correct to Python indexes
+                row -= 1
+                col -= 1
+                optsec = opts[secs[sec]]
+                val = {}
+                for key in spckeys:
+                    for alias in spckeys[key]:
+                        if alias in optsec:
+                            val[key] = optsec[alias]
+                            break
+                    else:
+                        val[key] = spcbase[key]
+                spcdat[row][col] = SpecLayout(**val)
+    for row in range(figdat['subp'][0]):
+        for col in range(figdat['subp'][1]):
+            if spcdat[row][col] is None:
+                spcdat[row][col] = SpecLayout(**spcbase)
+    # If axes merged, removed unnecessary labels
+    if figdat['shareaxes'] in ('Y', True) and ncols > 1:
+        for i in range(nrows):
+            for j in range(1, ncols):
+                spcdat[i][j].ylabel = None
+    if figdat['shareaxes'] in ('X', True) and nrows > 1:
+        for i in range(nrows-1):
+            for j in range(ncols):
+                spcdat[i][j].xlabel = None
 
     curves = {}
-    for sec in opts.sections():
-        if sec.strip().lower().startswith('curve'):
-            label, key = [item.strip() for item in sec.split(':', maxsplit=1)]
-            if label.lower() != 'curve':
+    for sec in secs:
+        if sec.startswith('curve'):
+            res = sec.split(':', maxsplit=1)
+            if res[0] != 'curve':
                 print(sec, 'will be ignored as a curve definition.')
                 continue  # This is not a right keyword, ignore.
-            curve = opts[sec]
-            if not curve.getboolean('show', fallback=True):
+            if len(res) != 2:
+                key = ' '
+            else:
+                key = secs[sec].split(':', maxsplit=1)[1].strip()
+            optsec = opts[secs[sec]]
+            # Check if curve to be shown
+            if not optsec.getboolean('show', fallback=True):
                 continue
-            curves[key] = {}
-            if 'file' not in curve:
+            # Subplot - check if subplot within range
+            res = optsec.get('subplot', fallback=None)
+            if res is not None:
+                row, col = parse_subid(res, figdat['subp'][1])
+                if row > figdat['subp'][0] or col > figdat['subp'][1]:
+                    continue
+                curves[key] = {'subplot': (row-1, col-1)}
+            else:
+                curves[key] = {'subplot': (-1, -1)}
+            if 'file' not in optsec:
                 print(f'WARNING: Missing file for "{sec}". Ignoring.')
                 continue
-            elif not os.path.exists(curve['file']):
+            elif not os.path.exists(optsec['file']):
                 fmt = 'ERROR: File "{}" not found in "{}".'
-                print(fmt.format(curve['file'], sec))
-            spc = curve.get('spectroscopy', fallback=None)
-            lvl = curve.get('level', fallback=None)
+                print(fmt.format(optsec['file'], sec))
+            spc = optsec.get('spectroscopy', fallback=None)
+            lvl = optsec.get('level', fallback=None)
             if spc is None or lvl is None:
                 raise ValueError('Spectroscopy not defined')
-            yid = curve.get('yaxis', None)
+            yid = optsec.get('yaxis', None)
             if yid is not None:
                 yid = 'y' + yid
-            curves[key]['data'] = Spectrum(curve['file'], spc, lvl, yid)
-            if curve.getboolean('broaden', fallback=False):
-                func = curve.get('function', None)
-                hwhm = curve.get('HWHM', None)
+            curves[key]['data'] = Spectrum(optsec['file'], spc, lvl, yid)
+            if optsec.getboolean('broaden', fallback=False):
+                func = optsec.get('function', None)
+                hwhm = optsec.get('hwhm', None)
                 if hwhm is not None:
                     hwhm = float(hwhm)
                 curves[key]['data'].set_broadening(hwhm, func, 'default')
             vizdata = {}
             for item in ('color', 'linestyle', 'linewidth'):
-                if curve.get(item, False):
-                    vizdata[item] = curve.get(item, False)
+                if optsec.get(item, False):
+                    vizdata[item] = optsec.get(item, False)
             if vizdata:
                 curves[key]['data'].set_display(**vizdata)
-            if curve.get('label', None) is not None:
-                curves[key]['data'].label = curve.get('label')
-            curves[key]['xshift'] = curve.getfloat('xshift', fallback=None)
-            curves[key]['xscale'] = curve.getfloat('xscale', fallback=None)
-            res = curve.get('yshift', None)
+            if optsec.get('label', None) is not None:
+                curves[key]['data'].label = optsec.get('label')
+            curves[key]['xshift'] = optsec.getfloat('xshift', fallback=None)
+            curves[key]['xscale'] = optsec.getfloat('xscale', fallback=None)
+            res = optsec.get('yshift', None)
             if res is not None:
                 try:
                     val = float(res)
@@ -195,9 +338,9 @@ def parse_inifile(fname) -> tp.Tuple[tp.Dict[str, tp.Any],
             else:
                 val = None
             curves[key]['yshift'] = val
-            curves[key]['yscale'] = curve.getfloat('yscale', fallback=None)
-            curves[key]['ynorm'] = curve.getboolean('normalize',
-                                                    fallback=False)
+            curves[key]['yscale'] = optsec.getfloat('yscale', fallback=None)
+            curves[key]['ynorm'] = optsec.getboolean('normalize',
+                                                     fallback=False)
 
     return figdat, spcdat, curves
 
@@ -219,43 +362,71 @@ def main() -> tp.NoReturn:
     else:
         figdata, spcdata, curves = parse_inifile(args.optfile)
 
-    fig, subp = plt.subplots(1, 1)
+    nrows, ncols = figdata['subp']
+    pars = {'tight_layout': True}
+    res = figdata['shareaxes']
+    if res == 'X' or res == True:
+        pars['sharex'] = True
+        if 'gridspec_kw' not in pars:
+            pars['gridspec_kw'] = {}
+        pars['gridspec_kw']['hspace'] = 0.0
+    if res == 'Y' or res == True:
+        pars['sharey'] = True
+        if 'gridspec_kw' not in pars:
+            pars['gridspec_kw'] = {}
+        pars['gridspec_kw']['wspace'] = 0.0
+    fig, subp = plt.subplots(nrows, ncols, **pars)
     if figdata['geom'] is not None:
         fig.set_size_inches(figdata['geom'])
-    for key in curves:
-        xaxis = np.array(curves[key]['data'].xaxis)
-        if curves[key]['xshift'] is not None:
-            xaxis += curves[key]['xshift']
-        if curves[key]['xscale'] is not None:
-            xaxis *= curves[key]['xscale']
-        yaxis = np.array(curves[key]['data'].yaxis)
-        if curves[key]['yshift'] is not None:
-            if curves[key]['yshift'] == 'base':
-                yshift = - np.min(yaxis)
+    for row in range(nrows):
+        for col in range(ncols):
+            if nrows > 1 and ncols > 1:
+                sub = subp[row, col]
+            elif nrows > 1:
+                sub = subp[row]
+            elif ncols > 1:
+                sub = subp[col]
             else:
-                yshift = curves[key]['yshift']
-            yaxis += yshift
-        if curves[key]['ynorm']:
-            yaxis /= np.max(np.abs(yaxis))
-        elif curves[key]['yscale'] is not None:
-            yaxis *= curves[key]['yscale']
-        stick = curves[key]['data'].get_broadening('func') == 'stick'
-        data = {}
-        if curves[key]['data'].label is not None:
-            data['label'] = curves[key]['data'].label
-        if curves[key]['data'].linecolor is not None:
-            data['color'] = curves[key]['data'].linecolor
-        if curves[key]['data'].linewidth is not None:
-            data['linewidth'] = curves[key]['data'].linewidth
-        if stick:
-            zeros = np.zeros(len(yaxis))
-            subp.vlines(xaxis, zeros, yaxis, **data)
-        else:
-            if curves[key]['data'].linestyle is not None:
-                data['linestyle'] = curves[key]['data'].linestyle
-            subp.plot(xaxis, yaxis, **data)
-        subp.legend()
-        spcdata.set_plot(subp)
+                sub = subp
+            for key in curves:
+                i, j = curves[key]['subplot']
+                if (i == row or i == -1) and (j == col or j == -1):
+                    xaxis = np.array(curves[key]['data'].xaxis)
+                    if curves[key]['xshift'] is not None:
+                        xaxis += curves[key]['xshift']
+                    if curves[key]['xscale'] is not None:
+                        xaxis *= curves[key]['xscale']
+                    yaxis = np.array(curves[key]['data'].yaxis)
+                    if curves[key]['yshift'] is not None:
+                        if curves[key]['yshift'] == 'base':
+                            yshift = - np.min(yaxis)
+                        else:
+                            yshift = curves[key]['yshift']
+                        yaxis += yshift
+                    if curves[key]['ynorm']:
+                        yaxis /= np.max(np.abs(yaxis))
+                    if curves[key]['yscale'] is not None:
+                        yaxis *= curves[key]['yscale']
+                    stick = curves[key]['data'].get_broadening('func') \
+                        == 'stick'
+                    data = {}
+                    if curves[key]['data'].label is not None:
+                        data['label'] = curves[key]['data'].label
+                    if curves[key]['data'].linecolor is not None:
+                        data['color'] = curves[key]['data'].linecolor
+                    if curves[key]['data'].linewidth is not None:
+                        data['linewidth'] = curves[key]['data'].linewidth
+                    if stick:
+                        zeros = np.zeros(len(yaxis))
+                        sub.vlines(xaxis, zeros, yaxis, **data)
+                    else:
+                        if curves[key]['data'].linestyle is not None:
+                            data['linestyle'] = curves[key]['data'].linestyle
+                        sub.plot(xaxis, yaxis, **data)
+                    sub.legend()
+                    spcdata[row][col].set_plot(sub)
+    if figdata['title'] is not None:
+        fig.suptitle(figdata['title'], fontweight='bold')
     plt.show()